Debugging elastic nl fwd double

+++ elastic_iso_lib_3d/elastic_iso_double/include/nonlinearPropElasticGpuFunctions_3D.h
@@ -5,7 +5,7 @@
 /*********************************** Initialization **************************************/
 // Gpu info
 bool getGpuInfo_3D(std::vector<int> gpuList, int info, int deviceNumber);
-void initNonlinearElasticGpu_3D(double dz, double dx, double dy, int nz, int nx, double ny, int nts, double dts, int sub, int minPad, int blockSize, double alphaCos, int nGpu, int iGpuId, int iGpuAlloc);
+void initNonlinearElasticGpu_3D(double dz, double dx, double dy, int nz, int nx, int ny, int nts, double dts, int sub, int minPad, int blockSize, double alphaCos, int nGpu, int iGpuId, int iGpuAlloc);
 void allocateNonlinearElasticGpu_3D(double *rhoxDtw, double *rhoyDtw, double *rhozDtw, double *lamb2MuDt, double *lambDtw, double *muxzDtw, double *muxyDtw, double *muyzDtw, int nx, int ny, int nz, int iGpu, int iGpuId);
 void deallocateNonlinearElasticGpu_3D(int iGpu, int iGpuId);
